+++ src/commands/functions.ts
@@ -11,7 +11,7 @@ const save = (vscode: any, fs: any, path: any, filename: string, content: string
   const pathfile = path.join(folder, filename);
 
   if (!fs.existsSync(path.dirname(pathfile))) {
-    fs.mkdirSync(path.dirname(pathfile));
+    fs.mkdirSync(path.dirname(pathfile), { recursive: true });
   }
 
   fs.access(pathfile, (err: any) => {