+++ src/extension.ts
@@ -6,12 +6,13 @@ import { cacheClear, cacheInfo } from './commands/cache';
 import command from './commands/command';
 import config from './commands/config';
 import controller from './commands/controller';
+import { dbCreate, dbSeed, dbTable } from './commands/db';
 import entity from './commands/entity';
 import filter from './commands/filter';
 import helper from './commands/helper';
 import key from './commands/key';
 import language from './commands/language';
-import { migrate, migrateRefresh, migrateRollback } from './commands/migrate';
+import { migrate, migrateRefresh, migrateRollback, migrateStatus } from './commands/migrate';
 import migration from './commands/migration';
 import model from './commands/model';
 import resource from './commands/resource';
@@ -66,6 +67,9 @@ export function activate(context: vscode.ExtensionContext) {
   const sparkTerminalMigrateRollback = vscode.commands.registerCommand('spark.terminal.migrate.rollback', () => {
     migrateRollback(vscode);
   });
+  const sparkTerminalMigrateStatus = vscode.commands.registerCommand('spark.terminal.migrate.status', () => {
+    migrateStatus(vscode);
+  });
   const sparkTerminalServe = vscode.commands.registerCommand('spark.terminal.serve', () => {
     serve(vscode);
   });
@@ -80,6 +84,15 @@ export function activate(context: vscode.ExtensionContext) {
   });
   const sparkTerminalRoutes = vscode.commands.registerCommand('spark.terminal.routes', () => {
     routes(vscode);
+  });
+  const sparkTerminalDbCreate = vscode.commands.registerCommand('spark.terminal.db.create', () => {
+    dbCreate(vscode);
+  });
+  const sparkTerminalDbSeed = vscode.commands.registerCommand('spark.terminal.db.seed', () => {
+    dbSeed(vscode);
+  });
+  const sparkTerminalDbTable = vscode.commands.registerCommand('spark.terminal.db.table', () => {
+    dbTable(vscode);
   });
 
   context.subscriptions.push(sparkFileCommand);
@@ -97,11 +110,15 @@ export function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(sparkTerminalMigrate);
   context.subscriptions.push(sparkTerminalMigrateRefresh);
   context.subscriptions.push(sparkTerminalMigrateRollback);
+  context.subscriptions.push(sparkTerminalMigrateStatus);
   context.subscriptions.push(sparkTerminalServe);
   context.subscriptions.push(sparkTerminalKey);
   context.subscriptions.push(sparkTerminalCacheClear);
   context.subscriptions.push(sparkTerminalCacheInfo);
   context.subscriptions.push(sparkTerminalRoutes);
+  context.subscriptions.push(sparkTerminalDbCreate);
+  context.subscriptions.push(sparkTerminalDbSeed);
+  context.subscriptions.push(sparkTerminalDbTable);
 }
 
 export function deactivate() {}