+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2023-09-03
+
+### Added
+
+- Add Database Commands
+
+### Fixed
+
+- Fix Language file creation
+
 ## [1.3.0] - 2023-08-31
 
 ### Added
@@ -46,7 +56,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.3.0...HEAD
+[unreleased]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/ManuelGil/vscode-codeigniter4-spark/compare/v1.0.1...v1.1.0